+++ x-pack/plugins/siem/common/detection_engine/schemas/response/__mocks__/utils.ts
