+++ x-pack/plugins/siem/common/detection_engine/schemas/response/error_schema.ts
