+++ x-pack/plugins/siem/common/detection_engine/schemas/response/find_rules_schema.ts
