+++ x-pack/plugins/siem/common/detection_engine/schemas/response/prepackaged_rules_schema.ts
