+++ x-pack/plugins/siem/common/detection_engine/schemas/response/schemas.ts
