+++ x-pack/plugins/siem/common/detection_engine/schemas/response/type_timeline_only_schema.ts
