+++ x-pack/plugins/siem/common/detection_engine/schemas/types/iso_date_string.test.ts
@@ -7,7 +7,7 @@
 import { IsoDateString } from './iso_date_string';
 import { pipe } from 'fp-ts/lib/pipeable';
 import { left } from 'fp-ts/lib/Either';
-import { getPaths, foldLeftRight } from '../../../../../../common/test_utils';
+import { foldLeftRight, getPaths } from '../../../test_utils';
 
 describe('ios_date_string', () => {
   test('it should validate a iso string', () => {
