+++ x-pack/plugins/siem/common/detection_engine/schemas/types/iso_date_string.ts
