+++ x-pack/plugins/siem/common/detection_engine/schemas/types/lists_default_array.test.ts
@@ -7,7 +7,7 @@
 import { ListsDefaultArray } from './lists_default_array';
 import { pipe } from 'fp-ts/lib/pipeable';
 import { left } from 'fp-ts/lib/Either';
-import { getPaths, foldLeftRight } from '../../../../../../common/test_utils';
+import { foldLeftRight, getPaths } from '../../../test_utils';
 
 describe('lists_default_array', () => {
   test('it should validate an empty array', () => {
