+++ x-pack/plugins/siem/common/detection_engine/schemas/types/lists_default_array.ts
