+++ x-pack/plugins/siem/common/detection_engine/schemas/types/positive_integer_greater_than_zero.ts
