+++ x-pack/plugins/siem/common/detection_engine/schemas/types/references_default_array.ts
