+++ x-pack/plugins/siem/common/detection_engine/schemas/types/uuid.test.ts
@@ -7,7 +7,7 @@
 import { UUID } from './uuid';
 import { pipe } from 'fp-ts/lib/pipeable';
 import { left } from 'fp-ts/lib/Either';
-import { foldLeftRight, getPaths } from '../../../../../../common/test_utils';
+import { foldLeftRight, getPaths } from '../../../test_utils';
 
 describe('uuid', () => {
   test('it should validate a uuid', () => {
