+++ x-pack/plugins/siem/server/lib/detection_engine/types.ts
@@ -4,10 +4,10 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { ListsDefaultArraySchema } from '../../../common/detection_engine/schemas/types/lists_default_array';
 import { CallAPIOptions } from '../../../../../../src/core/server';
 import { Filter } from '../../../../../../src/plugins/data/server';
 import { IRuleStatusAttributes } from './rules/types';
-import { ListsDefaultArraySchema } from './routes/schemas/types/lists_default_array';
 import { RuleAlertAction, RuleType } from '../../../common/detection_engine/types';
 
 export type PartialFilter = Partial<Filter>;
